pyproject.toml
[project]
name = "fastapi-secure-crud-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "passlib[argon2]>=1.7.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
